Remove cycle in printDiffs, diffLines and joinAlignedDiffs (#10818)
